Update Projectile.h
collisionInfo changed to return float type.

+++ CSC8503/Projectile.h
@@ -154,7 +154,7 @@ namespace NCL::CSC8503 {
 			
 		}
 		
-		 int collisionInfo() override{
+		 float collisionInfo() override{
 			return this->getExplosionRadius();
 		}
